#18 [RENAME] 폴더 구조 변경
