+++ src/pages/MyPage/DeleteAccountPage/DeleteAccountPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './DeleteAccountPage.module.css';
 import { Link } from 'react-router-dom';
-import Icon from '../../components/Icon/Icon.jsx';
+import Icon from '../../../components/Icon/Icon.jsx';
 
 const descriptions = [
   '• 아이디, 이메일, 학번은 .. 목적으로 6개월간 보관됩니다',
