+++ src/pages/MyPage/MyPage/MyPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styles from './MyPage.module.css';
 import { Link } from 'react-router-dom';
-import profileIcon from '../../assets/icon-profile.svg';
+import profileIcon from '../../../assets/icon-profile.svg';
 
 export default function MyPage() {
   return (
