+++ src/pages/MyPage/MyPage/MyPage.module.css
@@ -1,7 +1,7 @@
 .myPage {
   display: flex;
   flex-direction: column;
-  background-image: url('../../assets/mypage-background-gradation.svg');
+  background-image: url('../../../assets/mypage-background-gradation.svg');
   background-size: cover;
   background-position: center;
   width: 100%;
@@ -24,7 +24,7 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background-image: url('../../assets/logo-white.svg');
+  background-image: url('../../../assets/logo-white.svg');
   background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
