+++ src/pages/MyPage/index.js
@@ -1,0 +1,4 @@
+export { default as ChangePasswordPage } from './ChangePasswordPage/ChangePasswordPage';
+export { default as DeleteAccountPage } from './DeleteAccountPage/DeleteAccountPage';
+export { default as EditInfoPage } from './EditInfoPage/EditInfoPage';
+export { default as MyPage } from './MyPage/MyPage';
